[build] Merge Core Components as UI components.
First Stage in simplifing plexydesk, we merge Core Library
with UIKit library. this will help us to avoid duplicating
many of the classes in UIKit. and simplify the use of the new
PlexyDesk API. this will also allow us to directly use widgets, and
simplify plugins. and reduce the loading of plugins from other plugins.
Also helps us fix memory bugs and leaks in a single library.